requirements.txt
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
s3fs==2023.10.0
boto3==1.34.0
llama-index-core==0.10.0
llama-index-embeddings-openai==0.1.0
llama-index-storage-docstore-simple==0.1.0
llama-index-storage-kvstore-simple==0.1.0
llama-index-storage-index-store-simple==0.1.0 